test: split ledger reconciliation coverage
